Adds the ability to pause and resume threads, moved pause/resume code from PausableThreadPoolExecutor to internal code in the runTask method.
